Update App.js
Best practices to get two things from useState: the current state (count), and the function that lets you update it (setCount). You can give them any names, but the convention is to write [something, setSomething].

The first time the button is displayed, count will be 0 because passed 0 to useState(). When changing state, call setCount() and pass the new value to it. Clicking this button will increase the counter.